validação

+++ src/main/java/com/desafioCliente/cliente/services/ClientService.java
@@ -1,14 +1,20 @@
 package com.desafioCliente.cliente.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.desafioCliente.cliente.dto.ClientDTO;
 import com.desafioCliente.cliente.entities.Client;
 import com.desafioCliente.cliente.repositories.ClientRepository;
+import com.desafioCliente.cliente.services.exceptions.DatabaseException;
+import com.desafioCliente.cliente.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -19,8 +25,11 @@ public class ClientService {
 	private ClientRepository repository;
 	
 	@Transactional(readOnly = true)
+	
 	public ClientDTO findById(Long id) {	
-		Client client= repository.findById(id).get();	
+		
+		Client client= repository.findById(id).orElseThrow(
+				()-> new ResourceNotFoundException("Recurso não encontrado"));	
 		return new ClientDTO(client);
 		
 	}
@@ -42,17 +51,31 @@ public class ClientService {
 	
 	@Transactional
 	public ClientDTO update(Long id,ClientDTO dto) {
+		try {
 		Client entity =repository.getReferenceById(id);
 		copyDtoToEntity(dto,entity);
-
+		entity=repository.save(entity);
 		return new ClientDTO(entity);
 
 	}
-
-	@Transactional
+		catch(EntityNotFoundException e){
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
+	}
+	
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
 		repository.deleteById(id);
-
+		}
+    	catch (DataIntegrityViolationException e) {
+        	throw new DatabaseException("Falha de integridade referencial");
+   	}
+		
 	}
 	
 	private void copyDtoToEntity(ClientDTO dto, Client entity) {
